+++ text.h
@@ -16,6 +16,6 @@ public:
 
     void create(int size, string path);
     void display(SDL_Renderer* renderer, SDL_Color color, string text);
-    void render(SDL_Renderer* renderer, double x, double y, SDL_Rect* clip = NULL, double angle = 0, SDL_Point* center = NULL, SDL_RendererFlip flip = SDL_FLIP_NONE);
+    void render(SDL_Renderer* renderer, int x, int y, SDL_Rect* clip = NULL, double angle = 0, SDL_Point* center = NULL, SDL_RendererFlip flip = SDL_FLIP_NONE);
     void destroy();
 };
